Move `is_vector` to meta
